main.cpp initializes the UI and displays it. This is all main will need to do.
WindowLogin controls the login window, the first view the user interacts with upon opening the application

DatabaseConnection makes the connection to the MySQL server and is called from WindowLogin. Currently I am working on making the same server available for all of us.

Currently the only interaction event set is when the Sign In push_button is clicked: a window pops up as a testing tool to show that the database was in fact connected to. If the database was connected Searching... appears. Else: Invalid...

The QRC, or Resource File, consists of external resources loaded into the app. Righ now the only resource included is the EDI logo.

The .ui file is the XML file that contains the visual rules for the UI. This file is automaticall updated when changes are made in the Design module of QT and should not be edited directly.