Control Tab DB
This is the basic layout i created for the control tab. My goal was to make sure we would have the framework for mirroring the EDI datbase within the UI, so that. a manager or CEO can use the database. Currently, the data frin the MYSQL database from the employee table loads into the UI table. None of the buttons or search functions work yet.

In the View and Edit buttons, a window can pop up to show employee details in a more "User Friendly" visual style along with a calendar widget that represent the employee's schedule.

We could possibly combine these buttons as one, so that view and edit features can be done within the same window, considering that in order to edit, you have to view anyway!

Functions to edit cell data directly by clicking on a cell in the table will still need to be disabled, as unwanted errors can occur if it is that easy to edit an employee in the database.