Added Log Out Button
Will need someone to verify it since I dont have the MySQL DB setup. Also should we put the Log Out button on all the pages of the main tabWidget? Ex. since the Log Out button is on MyEDI page should we also have it on the Control and Analyze pages too?