QCalendar
Derived class that allows us to override how a QCalendarWidget object loads on the screen.

The class so far was only applied to the QCalendarWidget object in the controltab_vieweditemployee UI, for testing purposes. If we want to apply this class to the calnedar widget in the MyEDI tab, we must go to the windowmaincontainer.ui file, click on the calendar widget, select "Promote To" and add "QCalendar" as the promoted class.

A couple changes to the database must also be made in order for the date highlights to load properly. For instane the way the datanase is currently structured only allows for one scheudle to be associated with one employe, which is fine, but since a scheudle only consists of one shift at a time, we cannot associate more than one shift with one employee at a time. Shift and schedule should be split so that a schedule can contain multuple shifts.

As a short term solution, for testing purposes, in order to demo this part i had to remote the unique constraint on the employee attribute of the schedule table, so that I could associate multiple shifts with one person, to check that multiple shifts would load, highlighted, into the UI.

The date format of a Schedule must also be changed to 20201103 or yyyyMMdd.