Comments added in all files and shift highlights on MyEdi Schedule configured.
I commented all of the existing files (except for the Schedule and Shift classes) with a basic summary of each class's purpose of method use.

The QCalendarWidget in the MyEDI table now highlight the shifts, as it does in the Control Tab -> View/Edit Employee UI.

To achieve this, the windowmaincontainer class (which was previously configured with a QMainWindow base class), needed to be configured with a QWidget base class. As editing the .ui files became to messy to acheive this, I reworte the WindowMainContainer class as a sub QWidget class, from scratch. The class is now called WindowMain.