Employee and Team Table Searching, Add and View/Edit Teams/Projects
The tables for the Employees and the Teams are now searchable in the Contorl Tab.

A new team can be created by clicking "New Team" button.

An existing team's details can be viewed and modified by clicking on View/Edit Team button. Within the UI that pops up from this, various details can be viewed and modified such as

Team Members and Team Projects