QCalendar correct shift highlights and auto update of employee table on changes
The QCalendar widget shifts that are being highlighted on both calendars, in the MyEDI tab and in the View/Edit Employee UI, are now highlighting the correct shifts.

They were hard coded for testing purposes of the implentation before. Now the MyEdi calendar highlights shifts for the currently logged in user, and the View/Edit calendar highlights shifts for the currently selected employee to edit from the Employee table.

The Employee table will now automatically update when a user is in the View/Edit control tab and clickts so Save their Changes. Currently UPDATE statements are not being exexcuted, but when they are, the table will reflect the update statement.